Reverted to previous code with additions
Since the problem with the last couple days was the batteries and
not the code, this branch has reverted to a previous working version
of the code. Some of the first few changes were then re-added such as
adding MsTimer2 to startMode, so start mode changing into drive mode
works with this code. Scan is still to be added